% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET_documentation.R
\docType{data}
\name{DATASET}
\alias{DATASET}
\title{Sample Data for Analysis}
\format{
An object of class \code{data.frame} with 1000 rows and 3 columns.
}
\usage{
DATASET
}
\description{
A simple \code{data_frame} obtained by a Data Generating Process that is for testing and running of examples
\describe{
\item{y}{outcome variable}
\item{x}{two bivariate normal variables x1 and x2}
}
}
\keyword{datasets}
