% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_test.R
\name{lambda_test}
\alias{lambda_test}
\title{lambda_test}
\usage{
lambda_test(f, nB)
}
\arguments{
\item{f}{An "iOLS_path" fitted model object that you want to apply this test on.}

\item{nB}{The number of iteration that you want to be done in the bootstrap process used in the function.}
}
\value{
a lambda_test object
}
\description{
Printing and plotting of the lambda-test.
}
