% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iOLS.R, R/plot_iOLS_path.R
\name{plot}
\alias{plot}
\alias{plot.iOLS}
\alias{plot.iOLS_path}
\title{plot.iOLS}
\usage{
plot(m, plot_beta, ...)

\method{plot}{iOLS}(m, ..., plot_beta = "")

plot(m, plot_beta, ...)

\method{plot}{iOLS_path}(m, delta_rank = NULL, plot_beta = "", ...)
}
\arguments{
\item{m}{An "iOLS_path" fitted model object.}

\item{plot_beta}{If you want to see the trajectory of one estimated parameter beta only, just precise plot_beta = k (k=0 if you want to see the intercept's trajectory for example). Otherwise, write plot_beta = "" (the default value), and you will see all parameters' trajectory. In this case, the colors of each curve is assigned randomly, but by precising which parameters' trajectory you want to see, it will be drawn in black.}

\item{...}{other parameters.}

\item{delta_rank}{Among all the hyper-parameters delta, we can choose to plot the "iOLS_path" fitted model object corresponding to the chosen delta_rank. If a value is not precised, the default value is NULL and the function will only display the estimated parameter(s) in function of log(delta).}
}
\value{
a plot of a iOLS fitted model object.

a plot of a iOLS_path fitted model object.
}
\description{
See the notebook at: https://www.davidbenatia.com/.

See the notebook at: https://www.davidbenatia.com/.
}
