% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iOLS_print.R
\name{print.iOLS}
\alias{print.iOLS}
\title{print.iOLS}
\usage{
\method{print}{iOLS}(m, ...)
}
\arguments{
\item{m}{An \code{iOLS} fitted model object.}

\item{...}{other parameters.}
}
\value{
a display of an \code{iOLS} fitted model object.
}
\description{
Function that prints an \code{iOLS} fitted model object.
}
\examples{
data(DATASET)
y = DATASET$y
x = as.matrix(DATASET[,c("X1","X2")])
lm = lm(log(y+1) ~ x)
lm_coef = c(coef(lm))
X = cbind(rep(1, nrow(x)), x)
tX = t(X)
library(matlib) ; VX = inv(tX \%*\% X)
f = iOLS(y, X, VX, tX, 20, b_init = lm_coef)
print(f)
}
