% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_test_print.R
\name{print.lambda_test}
\alias{print.lambda_test}
\title{print.lambda_test#'}
\usage{
\method{print}{lambda_test}(m, ...)
}
\arguments{
\item{m}{A \code{lambda_test} object.}

\item{...}{other parameters.}
}
\value{
a display and a plot of a \code{lambda_test} object.
}
\description{
Function that prints a \code{lambda_test} object.
}
\examples{
data(DATASET)
y = DATASET$y
x = as.matrix(DATASET[,c("X1","X2")])
lm = lm(log(y+1) ~ x)
lm_coef = c(coef(lm))
X = cbind(rep(1, nrow(x)), x)
k = iOLS_path(y, X, b_init = lm_coef, deltainf = 10^-5,
deltasup = 10^4, nbre_delta = 20,
epsi = 10^-3, error_type = "HC0")

L = lambda_test(k, nB = 5)

print(L)
}
