% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Lambda_test.R, R/print_iOLS.R,
%   R/print_iOLS_path.R
\name{print}
\alias{print}
\alias{print.lambda_test}
\alias{print.iOLS}
\alias{print.iOLS_path}
\title{print.lambda_test}
\usage{
print(m, ...)

\method{print}{lambda_test}(m, ...)

print(m, ...)

\method{print}{iOLS}(m, ...)

print(m, ...)

\method{print}{iOLS_path}(m, delta_rank = NULL, ...)
}
\arguments{
\item{m}{An "iOLS_path" fitted model object.}

\item{...}{other parameters.}

\item{delta_rank}{Among all the hyper-parameters delta, we can choose to plot the "iOLS_path" fitted model object corresponding to the chosen delta_rank. If a value is not precised, the default value is NULL and the function will only display the estimated parameter(s) in function of log(delta).}
}
\value{
a display and a plot of a lambda_test object.

a display of a iOLS fitted model object.

a display of a iOLS_path fitted model object.
}
\description{
See the notebook at: https://www.davidbenatia.com/.

See the notebook at: https://www.davidbenatia.com/.

See the notebook at: https://www.davidbenatia.com/.
}
